+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## master (unreleased)
 
+### New Features
+
+* Add support of parsing `PivotCacheDefinition#cacheFields`
+* Add support of parsing `CacheFields#cache_field`
+* Add support of parsing `CacheField#shared_item`
+
 ### Changes
 
 * Changes from `rubocop` v1.4.0